moves around new layout components into more coherent location
