+++ web-local/src/lib/components/layouts/index.ts
@@ -1,2 +1,1 @@
-export { initializeNodeStoreContexts } from "./initialize-node-store-contexts";
 export { default as RillDeveloperApplicationLayout } from "./RillDeveloperApplicationLayout.svelte";
