+++ web-local/src/routes/(application)/+layout.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import { RillDeveloperApplicationLayout } from "@rilldata/web-local/lib/layouts";
+  import { RillDeveloperApplicationLayout } from "@rilldata/web-local/lib/components/layouts";
 </script>
 
 <RillDeveloperApplicationLayout>
