+++ web-local/src/routes/+layout.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import { initializeNodeStoreContexts } from "@rilldata/web-local/lib/layouts";
+  import { initializeNodeStoreContexts } from "@rilldata/web-local/lib/application-state-stores/initialize-node-store-contexts";
   /** This function will initialize the existing node stores and will connect them
    * to the Node server.
    */
